RawValueInliner split into several classes
